extract App, runners, and RunState from cli.tsx into testable modules
